Create dedicated admin file
